Add sign-in, sign-out, and sign-up
User must confirm password upon sign-up. Password must be longer than 8 characters. A user can decide if he or she wants the app to remember them when signing in. If password is forgotten, an email will be sent with a link to reset the password. Password changes persist in the database.