Add functionality to sign-in and sign-up buttons in nav bar
On the landing page, a user can click on the sign in button if he or she has an account.
If the user does not have an account, the user can click on the sign-up button.
The buttons will take the user to the correct corresponding pages, where the data they enter persists into the database.