Add styling to homepage.
The nav bar is mostly functional with the exception that once users enter their credential for sign-up, they are not being redirected.
In addition, the forgot password email function needs to be tweaked.

I am moving away from this aspect of the application to focus more on functionality and to implement desired features.