Update link names on navigation bar
The navigation bar links are now relevant to the app. When the user is logged in, a drop down menu with their username on it shows up. This menu has two functional links: one that allows the user to edit his or her account and another that directs the user to his or her profile page.